Recommend docker-ce instead of docker.io
Kubernetes seems to work better with the version of Docker packaged by
docker.com, so we want to use that version.  Installation requires a
few steps so I'll remove docker.io from the set of packages that get
installed by init-dependencies and add a note to the README pointing
to the docker.com installation web page.