Run docker as $USER, not root
We want to run docker as the current unprivileged user, not as root.
If you can't run docker commands, you probably need to add your user
to the `docker` group and log out and back in.  When you run `id` you
should see that you're a member of the `docker` group.