Use Docker multistage builds to build prod images
We're now building the executables for the docker images inside of
docker instead of building them externally and copying them into the
docker image.  This will be more repeatable since we'll be able to
build on CI machines that don't have a development environment set up.

The prod Dockerfile is now a 3-stage build and uses Docker multistage
explicitly.  We no longer need an ad-hoc "system" image.