Install libbpf headers outside of git submodule
We used to install the libbpf header files into libbpf/src/usr/include
but because this wasn't ignored by git the libbpf submodule was always
"dirty". Now we install them into headers/bpf and tell git to ignore
that directory.

I also made building libbpf more explicit by adding it to the list of
subdirectories in the top-level Makefile. It was implicitly built by
common/common.mk but that was tricky to figure out. This is more
obvious.