Refresh bpf_elf.h from iproute2
The motivation for this change was us wanting to add BTF metadata to
our maps, and to do that we need the BPF_ANNOTATE_KV_PAIR macro. From
what I can tell, the canonical version of bpf_elf.h is the one in the
iproute2 project so I'll get it from there.

I tried to update bpf_util.h, too, but that caused our compiles to
fail so that will have to be a project for another day.