DRY up CLI code by moving duplicated code into a common module
There was a *lot* of cut-n-paste in the CLI code, this reduces it.

There was a framework for doing this in the xdp-tutorial (which is the
baseline for this project) but for some reason it wasn't used. I added
a common/cli_util.c/h and had the other CLI programs reference it. I
didn't have time to check all of the programs for duplicated code,
though, so there's probably more.