Add tunnelID to GUE ping packet
We hit a snag with the GUE ping packet because we were trying to use
one value for two purposes. We were sending one ID in the packet but
then trying to use it to both update the tunnel endpoints on the EPIC
side (which is the job of gid-sid) and also checking the
password (which is the job of the tunnelID).

For the time being we're going to send both the gid-sid and the
tunnelID in the ping packet. We hope that this isn't a long-term
solution because we'd prefer to have the tunnelID be local to EPIC but
since we're already sending it to PureLB we might as well use it to
get ourselves out of this corner in which we've painted ourselves.

As part of this I removed the "gue->hlen == 1" ping since it can't
possibly work anymore. From now on pings are required to have the
password.

I opted for caution and didn't add the tunnelID to *all* of the GUE
headers (i.e., gueexthdr). I made a new data structure just for
pings. We might need to have everything send tunnelID but that will
take a non-trivial effort. I wanted to get this working.