Don't redirect decapped packets on the EPIC unless CFG_RX_FWD is set
In most cases we don't need to redirect packets after we decap them on
the EPIC because we're on the input side of the interface so Linux
will figure out what to do with them after we let them go.

In some cases this forwarding is benign, but in other cases it can
actually be harmful.

Just in case we need this functionality someday I've hidden it behind
the CFG_RX_FWD config flag which is the same one we use for the same
purpose on the PureLB side.