Add a del-service cmd to cli_service
Our new approach to configuring the PFC is to have the rep and pod
controllers maintain data structures in the LoadBalancer CR which are
then applied by the LB agent controllers. Those controllers loop
through an entire service at a time so we can delete the whole service
then re-create it on each pass.

There's a very small race condition where packets that arrive between
the delete and the re-build can get lost but we can optimize that
later.