Trace only if it's our packet
We used to output about 4 lines in the trace log even if the packet
wasn't one of ours, but it's annoying to have to sift through all of
the decap chaff when you're looking for an encap bug or vice
versa. Now we log nothing until we have decided that the packet is one
of ours.