Enhance parse_headers() to also parse the basic GUE header
We've been having trouble with too-short packets and one theory is
that we've stuffed up the pointer arithmetic between parse_headers and
the decap code. This change modifies the headers struct so it includes
an explicit gue header pointer and offset. It also adds GUE header
parsing code to parse_headers to fill in that field.

It also changes the decap function to always use the pointer in the
header struct instead of copying it.