Use bpf_skb_pull_data() to make packet headers linear
We'd like to parse all of the headers at once, but sometimes the
packet's skb is "non-linear" and doesn't have the full
headers (including GUE) in the data field. It always seems to have up
to the TCP/UDP header but we need to use bpf_skb_pull_data() to make
sure that we have the GUE header also. This lets us use direct access
to parse the headers.