"Linearize" packets after we decap them
In earlier commits we learned about the subtle magic of
bpf_skb_pull_data() but we didn't apply what we learned in enough
places. It turns out that the same thing that happens with
encapsulated packets can happen after we decapsulate them so we need
to check and maybe pull after gue_decap_v4().