Switch map_decap key from local endpoint to remote endpoint
When we tried to deploy client clusters on AWS we hit a snag: AWS's
CNI source-NATs port values so when we sent GUE packets on port 6080
the CNI would munge that to a random port value. Not cool, Amazon.

Because EPIC runs in an environment that we control, and the client
cluster doesn't, it's better to use values that come from EPIC as keys
where possible. In this case map_decap is just a lightweight check to
filter out non-GUE packets. If the packet matches an entry in
map_decap it doesn't mean that it's a GUE packet, but if there's no
match then it definitely isn't.