Decap based on either map_decap or dest port
AWS nodes can send traffic back to EPIC via a different route than the
GUE ping packets, so we can't necessarily use map_decap to tell
whether a packet is GUE or something else. This change adds a second
check so if the map_decap lookup fails we'll still try to decap the
packet if the dest port is 6080.

This is a much smaller change than the diff makes it appear because a
big chunk of code was re-indented but no other changes were made.