Remove debug flag from BPF compile
I ran into this problem[1] which appears to be caused by linux kernel
version skew between the host that compiles the filters and the one
that loads them. I'd like to be able to compile on my machine and load
into Ubuntu 22.04 hosts. Compiling without "-g" increases the version
compatibility so let's remove that flag for now.

[1] https://github.com/xdp-project/xdp-tutorial/issues/38