Always use "0" for the port in the encap table key
The encap table is both very large and very active and this might be
causing problems. This is an experiment to make the table smaller (by
thousands of entries) by not tracking the source port anymore.

The thinking is that we would never need to encap a packet going to
10.42.2.2:12345 using tunnel 59 but also encap a packet going to
10.42.2.2:54321 using tunnel 27. If it came from a pod then it's going
back to that pod.